+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/listener/http2/Http2SourceConnectionHandlerBuilder.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.wso2.transport.http.netty.listener;
+package org.wso2.transport.http.netty.listener.http2;
 
 
 import io.netty.handler.codec.http2.AbstractHttp2ConnectionHandlerBuilder;
@@ -28,6 +28,7 @@ import io.netty.handler.codec.http2.Http2Settings;
 import org.wso2.transport.http.netty.common.Constants;
 import org.wso2.transport.http.netty.common.FrameLogger;
 import org.wso2.transport.http.netty.contract.ServerConnectorFuture;
+import org.wso2.transport.http.netty.listener.HttpServerChannelInitializer;
 
 import static io.netty.handler.logging.LogLevel.TRACE;
