+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/listener/http2/Http2WithPriorKnowledgeHandler.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.transport.http.netty.listener;
+package org.wso2.transport.http.netty.listener.http2;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufUtil;
@@ -26,6 +26,7 @@ import io.netty.channel.ChannelPipeline;
 import io.netty.handler.codec.http2.Http2CodecUtil;
 import org.wso2.transport.http.netty.common.Constants;
 import org.wso2.transport.http.netty.contract.ServerConnectorFuture;
+import org.wso2.transport.http.netty.listener.HttpServerChannelInitializer;
 
 import static org.wso2.transport.http.netty.common.Util.safelyRemoveHandlers;
 import static java.lang.Math.min;
